requirements.txt
Django==3.2.18
django-extensions
psycopg2
ipython
Pillow
djangorestframework
markdown
django-filter